+++ intercom_flutter_platform_interface/test/method_channel_intercom_flutter_test.dart
@@ -455,6 +455,30 @@ void main() {
         <Matcher>[isMethodCall('fetchLoggedInUserAttributes', arguments: null)],
       );
     });
+
+    test('setUserJwt', () async {
+      await intercom.setUserJwt('test');
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('setUserJwt', arguments: {
+            'jwt': 'test',
+          })
+        ],
+      );
+    });
+
+    test('setAuthTokens', () async {
+      await intercom.setAuthTokens({'security_token': 'test'});
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('setAuthTokens', arguments: {
+            'tokens': {'security_token': 'test'},
+          })
+        ],
+      );
+    });
   });
 }