+++ intercom_flutter_web/example/integration_test/intercom_flutter_web_test.dart
@@ -146,5 +146,13 @@ void main() {
     testWidgets('fetchLoggedInUserAttributes', (WidgetTester _) async {
       expect(plugin.fetchLoggedInUserAttributes(), completes);
     });
+
+    testWidgets('setUserJwt', (WidgetTester _) async {
+      expect(plugin.setUserJwt('test'), completes);
+    });
+
+    testWidgets('setAuthTokens', (WidgetTester _) async {
+      expect(plugin.setAuthTokens({'security_token': 'test'}), completes);
+    });
   });
 }